Separated the hugepage handler into unspported and linux
