--- pkg/kubelet/cm/hugepagehandler/fake_hugepage_handler.go
@@ -1,39 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package hugepagehandler
-
-import (
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/klog"
-)
-
-type fakeHandler struct {
-}
-
-func (m *fakeHandler) Start(containerRuntime runtimeService) {
-	klog.Info("[fake hugepagehandler] Start()")
-}
-
-func (m *fakeHandler) AddContainer(pod *v1.Pod, container *v1.Container, containerID string) error {
-	klog.Infof("[fake hugepagehandler] AddContainer (pod: %s, container: %s, container id: %s)", pod.Name, container.Name, containerID)
-	return nil
-}
-
-// NewFakeHandler create fake hugepage handler
-func NewFakeHandler() Handler {
-	return &fakeHandler{}
-}
